+++ DarmuhsTerminalCommands/EventSub/TerminalGeneral.cs
@@ -14,7 +14,7 @@ namespace TerminalStuff.EventSub
             //Plugin.instance.Config.Reload();
             MenuBuild.ClearMyMenustuff();
             ConfigSettings.TerminalStuffMain.DeleteAll();
-            lastNode = null; //stop persisting nodes between lobbies
+            lastNode.displayText = ""; //stop persisting nodes between lobbies
             lastText = "";
             //Plugin.ClearLists();
             //Terminal disabled, disabling ESC key listener OnDisable