+++ DarmuhsTerminalCommands/PluginCore/Plugin.cs
@@ -33,7 +33,7 @@ namespace TerminalStuff
         {
             public const string PLUGIN_GUID = "darmuh.TerminalStuff";
             public const string PLUGIN_NAME = "darmuhsTerminalStuff";
-            public const string PLUGIN_VERSION = "3.5.1";
+            public const string PLUGIN_VERSION = "3.5.4";
         }
 
         internal static ManualLogSource Log;